+++ pumpkin/src/entity/player.rs
@@ -85,8 +85,8 @@ use crate::net::{ClientPlatform, GameProfile};
 use crate::net::{DisconnectReason, PlayerConfig};
 use crate::plugin::player::player_change_world::PlayerChangeWorldEvent;
 use crate::plugin::player::player_gamemode_change::PlayerGamemodeChangeEvent;
-use crate::plugin::player::player_kick::PlayerKickEvent;
-use crate::plugin::player::player_level_change::PlayerLevelChangeEvent;
+use crate::plugin::player::kick::PlayerKickEvent;
+use crate::plugin::player::level_change::PlayerLevelChangeEvent;
 use crate::plugin::player::player_teleport::PlayerTeleportEvent;
 use crate::server::Server;
 use crate::world::World;
