+++ pumpkin/src/plugin/api/events/player/mod.rs
@@ -7,9 +7,9 @@ pub mod player_interact_entity_event;
 pub mod player_interact_event;
 pub mod player_interact_unknown_entity_event;
 pub mod player_join;
-pub mod player_kick;
+pub mod kick;
 pub mod player_leave;
-pub mod player_level_change;
+pub mod level_change;
 pub mod player_login;
 pub mod player_move;
 pub mod player_teleport;
